localization for js

+++ public/js/product.js
@@ -1,7 +1,7 @@
 
 function deleteProduct(e, id) {
   e.preventDefault();
-  msg = "Do you want to delete Product ID = " + id;
+  msg = Lang.get('product.index.delete_confirm') + id;
   if (confirm(msg)) {
     document.getElementById('delete-prd'+id).submit();
   }
@@ -9,15 +9,17 @@ function deleteProduct(e, id) {
 
 function deleteImage(url, e) {
   e.preventDefault();
-  $.ajax({
-    headers: {
-      'X-CSRF-TOKEN': $('#csrf-token').attr('content')
-    },
-    type: "delete",
-    url: url,
-    data: {},
-    success: function (result) {
-      document.getElementById('img-' + result.id).remove();
-    }
-  });
+  if(confirm(Lang.get('product.update.delete_confirm'))){
+    $.ajax({
+      headers: {
+        'X-CSRF-TOKEN': $('#csrf-token').attr('content')
+      },
+      type: "delete",
+      url: url,
+      data: {},
+      success: function (result) {
+        document.getElementById('img-' + result.id).remove();
+      }
+    });
+  }
 }
